Refactor MRSRequest forms
One class name (MRSRequestForm) was defined twice, another was not used
anymore but kept in case (MRSRequestAdminForm), and the first
MRSRequestForm had part of its feature in the MRSRequestCreateForm so
both were merged.

Old test for MRSRequestAdminForm rewritten to nicely cover the
attachment part from validation to relation save in obscure django
conditions (with commit=False).

Proactive tech debt payment in prevision for a new, and fully custom,
updateform for some of the staff, pretty much the warmup needed for
the next commit that has manual cooking which requires
form.save(commit=False) ;)