Sentry sourcemap release + introduce predeploy hook
So, there's one step forward having webpack bundle tracebacks unpacked
with the sourcemaps, + a refactor / new feature of the deployment.

The new feature is that playlabs will execute /app/predeploy script (if
any) before running the actual server, in the same context but in a
temporary container, in the fashion it was running migrate.

So, in playlabs, migrate step was removed in favor of /app/predeploy.
In clilabs, automation for sentry sourcemap upload was added.
Finnaly, betagouv/mrs's app/predeploy calls the above to upload the
sourcemaps at the right moment.