Bugfix formset js
Before using FormSet, the name of the first date fields were date_depart
and date_return, and they were not affected by this code that is fired
by the iterative checkbox and that used to reset all date fields that
were below, and that were all starting with a number ie. 1-date_return

It was formsets without formsets because we had parts above and below
the iterative checkbox.