Fix #849: Alphanumeric (corsican) NIR support
Migrate Person.nir from BigInteger to CharField.

This makes the form-level override not-necessary anymore, allowing to
rely purely on model-level validators.

Test all sorts of combos to be as helpful as possible for the user in
case of NIR validation failure.

Also introduces a .po locale file to change strings, as requested by our
dear developer @vindarel, which means that the Dockerfile now compiles
translations.