Filter out imported requests from monthly email exports
Also add transport date to imported test MRSRequest fixture
201805020002.

MRSRequest 201805020002 should not appear in monthly email because it
has a mandate date which means that it has been imported.

As such, it was expected for the developer to see this MRSRequest *not*
appear in the test result.

However, it didn't appear in the test result for another reason: it had
no transport date. As such, it received the treatment of pre-MRS legacy
data which is to be skipped.

201805020002 was skipped which passed test indeed, but not for the
appropriate reason.

As such, this patch adds transport dates to 201805020002, which of
course updates the resulting detail page fixture, but also some other
stat tests.