Making sense out of local admin updates
Initially trying to hide the password reset view from supervisors
because it's defunkt, by simply moving the checks into the View.has_perm
method that's used to generate menus based on per-object permissions,
instead of the form that still displays and says that it's not going to
work.

It seems a bit worse than I thought, for some reason new code was built
on the legacy profile attribute which dates from *before* users could
have multiple groups.